+++ cggmp21-keygen/src/non_threshold.rs
@@ -66,10 +66,10 @@ pub struct MsgRound2<E: Curve, L: SecurityLevel> {
     /// $A_i$
     pub sch_commit: schnorr_pok::Commit<E>,
     /// Party contribution to chain code
-    #[cfg(feature = "hd-wallets")]
+    #[cfg(feature = "hd-wallet")]
     #[serde_as(as = "Option<utils::HexOrBin>")]
     #[udigest(as = Option<udigest::Bytes>)]
-    pub chain_code: Option<slip_10::ChainCode>,
+    pub chain_code: Option<hd_wallet::ChainCode>,
     /// $u_i$
     #[serde(with = "hex::serde")]
     #[udigest(as_bytes)]
@@ -127,7 +127,7 @@ pub async fn run_keygen<E, R, M, L, D>(
     sid: ExecutionId<'_>,
     rng: &mut R,
     party: M,
-    #[cfg(feature = "hd-wallets")] hd_enabled: bool,
+    #[cfg(feature = "hd-wallet")] hd_enabled: bool,
 ) -> Result<CoreKeyShare<E>, KeygenError>
 where
     E: Curve,
@@ -159,9 +159,9 @@ where
     let mut rid = L::Rid::default();
     rng.fill_bytes(rid.as_mut());
 
-    #[cfg(feature = "hd-wallets")]
+    #[cfg(feature = "hd-wallet")]
     let chain_code_local = if hd_enabled {
-        let mut chain_code = slip_10::ChainCode::default();
+        let mut chain_code = hd_wallet::ChainCode::default();
         rng.fill_bytes(&mut chain_code);
         Some(chain_code)
     } else {
@@ -176,7 +176,7 @@ where
         rid,
         X: X_i,
         sch_commit,
-        #[cfg(feature = "hd-wallets")]
+        #[cfg(feature = "hd-wallet")]
         chain_code: chain_code_local,
         decommit: {
             let mut nonce = L::Rid::default();
@@ -278,7 +278,7 @@ where
         return Err(KeygenAborted::InvalidDecommitment(blame).into());
     }
 
-    #[cfg(feature = "hd-wallets")]
+    #[cfg(feature = "hd-wallet")]
     let chain_code = if hd_enabled {
         tracer.stage("Calculate chain_code");
         let blame = utils::collect_simple_blame(&decommitments, |decom| decom.chain_code.is_none());
@@ -286,7 +286,7 @@ where
             return Err(KeygenAborted::MissingChainCode(blame).into());
         }
         Some(decommitments.iter_including_me(&my_decommitment).try_fold(
-            slip_10::ChainCode::default(),
+            hd_wallet::ChainCode::default(),
             |acc, decom| {
                 Ok::<_, Bug>(utils::xor_array(
                     acc,
@@ -366,7 +366,7 @@ where
                 .map(|d| d.X)
                 .collect(),
             vss_setup: None,
-            #[cfg(feature = "hd-wallets")]
+            #[cfg(feature = "hd-wallet")]
             chain_code,
         },
         x: x_i,