+++ cggmp21/src/lib.rs
@@ -300,8 +300,8 @@
 #![cfg_attr(not(test), forbid(unused_crate_dependencies))]
 #![cfg_attr(docsrs, feature(doc_auto_cfg))]
 
-#[cfg(feature = "hd-wallets")]
-pub use slip_10;
+#[cfg(feature = "hd-wallet")]
+pub use hd_wallet;
 pub use {
     generic_ec, paillier_zk,
     paillier_zk::{fast_paillier, rug},